+++ gui/src/app/pages/NwbPage/NwbMainView/DefaultNwbFileView.tsx
@@ -25,6 +25,8 @@ const DefaultNwbFileView: FunctionComponent<Props> = ({
 }) => {
   const rootGroup = useGroup(nwbFile, "/");
   const processingGroup = useGroup(nwbFile, "/processing");
+  const includeStimulus = useIncludeStimulus(nwbFile);
+  const includeUnits = useIncludeUnits(nwbFile);
   const headings = useMemo(() => {
     const hh: Heading[] = [];
     hh.push({
@@ -47,15 +49,36 @@ const DefaultNwbFileView: FunctionComponent<Props> = ({
         groupPath: "",
       });
     }
-    hh.push({
-      name: "units",
-      label: "units",
-      groupPath: "/units",
-    });
+    if (includeUnits) {
+      hh.push({
+        name: "units",
+        label: "units",
+        groupPath: "/units",
+      });
+    }
     if (rootGroup) {
       for (const sg of rootGroup.subgroups) {
         if (sg.name === "processing") {
+          // we already took care of processing above
           continue;
+        }
+        if (sg.name === "specifications") {
+          // do not include specifications in the top level headings
+          continue;
+        }
+        if (sg.name === "general") {
+          // do not include general in the top level headings
+          continue;
+        }
+        if (sg.name === "units") {
+          // already handled units above
+          continue;
+        }
+        if (sg.name === "stimulus") {
+          // only include stimulus if there is something in it besides empty groups
+          if (!includeStimulus) {
+            continue;
+          }
         }
         if (!hh.find((h) => h.groupPath === sg.path)) {
           hh.push({
@@ -72,7 +95,7 @@ const DefaultNwbFileView: FunctionComponent<Props> = ({
       return 0;
     });
     return hh;
-  }, [processingGroup, rootGroup]);
+  }, [processingGroup, rootGroup, includeStimulus, includeUnits]);
   return (
     <div style={{ position: "absolute", width, height, overflowY: "auto" }}>
       {/* <DendroLinksView /> */}
@@ -86,6 +109,81 @@ const DefaultNwbFileView: FunctionComponent<Props> = ({
       ))}
     </div>
   );
+};
+
+const useIncludeStimulus = (nwbFile: RemoteH5FileX) => {
+  const [includeStimulus, setIncludeStimulus] = useState<boolean | undefined>(
+    undefined,
+  );
+  useEffect(() => {
+    let canceled = false;
+    const load = async () => {
+      const group = await nwbFile.getGroup("/stimulus");
+      if (canceled) return;
+      if (!group) {
+        setIncludeStimulus(false);
+        return;
+      }
+      let foundSomething = false;
+      for (const sg of group.subgroups) {
+        const subgroup = await nwbFile.getGroup(sg.path);
+        if (
+          subgroup &&
+          (subgroup.subgroups.length > 0 || subgroup.datasets.length > 0)
+        ) {
+          foundSomething = true;
+          break;
+        }
+      }
+      if (canceled) return;
+      setIncludeStimulus(foundSomething);
+    };
+    load();
+    return () => {
+      canceled = true;
+    };
+  }, [nwbFile]);
+  return includeStimulus;
+};
+
+const useIncludeUnits = (nwbFile: RemoteH5FileX) => {
+  const [includeUnits, setIncludeUnits] = useState<boolean | undefined>(
+    undefined,
+  );
+  useEffect(() => {
+    let canceled = false;
+    const load = async () => {
+      const group = await nwbFile.getGroup("/units");
+      if (canceled) return;
+      if (!group) {
+        setIncludeUnits(false);
+        return;
+      }
+      if (group.datasets.length > 0) {
+        if (canceled) return;
+        setIncludeUnits(true);
+        return;
+      }
+      let foundSomething = false;
+      for (const sg of group.subgroups) {
+        const subgroup = await nwbFile.getGroup(sg.path);
+        if (
+          subgroup &&
+          (subgroup.subgroups.length > 0 || subgroup.datasets.length > 0)
+        ) {
+          foundSomething = true;
+          break;
+        }
+      }
+      if (canceled) return;
+      setIncludeUnits(foundSomething);
+    };
+    load();
+    return () => {
+      canceled = true;
+    };
+  }, [nwbFile]);
+  return includeUnits;
 };
 
 type TopLevelHeadingViewProps = {
@@ -109,7 +207,7 @@ const TopLevelHeadingView: FunctionComponent<TopLevelHeadingViewProps> = ({
     group &&
     group.subgroups.length + group.datasets.length > 0;
   return (
-    <div style={{ marginLeft: 10 }}>
+    <div style={{ marginLeft: 10, userSelect: "none" }}>
       <div
         style={{
           cursor: "pointer",